+++ src/system/extensions/index.ts
@@ -1,4 +1,5 @@
-import './Array';
-import './Date';
-import './String';
-import './Number';
+import "./Array";
+import "./Date";
+import "./String";
+import "./Number";
+import "./Boolean";