+++ test/system/extensions/Boolean.ts
@@ -1,0 +1,5 @@
+import "../../../src/system/extensions/index";
+
+const isRight: boolean | undefined = false;
+
+console.log(Boolean.f_hasValue(isRight));