+++ test/system/extensions/String.ts
@@ -1,4 +1,4 @@
-import '../../../src/system/extensions/index';
+import "../../../src/system/extensions/index";
 
 console.log("f_toLower", "CN.Troy&鱼摆摆".f_toLower() === "cn.troy&鱼摆摆");
 
@@ -8,7 +8,7 @@ console.log("f_isEmpty", !"鱼摆摆".f_isEmpty());
 
 console.log("f_isEmpty", "".f_isEmpty());
 
-console.log("f_insert", "鱼摆摆".f_insert("cn.troy", 2) === "鱼摆cn.troy摆")
+console.log("f_insert", "鱼摆摆".f_insert("cn.troy", 2) === "鱼摆cn.troy摆");
 
 console.log("f_contains", !"鱼摆摆,cn.troy".f_contains("BrendanEich"));
 console.log("f_contains", "鱼摆摆,cn.troy".f_contains("摆摆"));
@@ -16,19 +16,31 @@ console.log("f_contains", "鱼摆摆,cn.troy".f_contains("摆摆"));
 console.log("f_remove", "鱼摆摆,cn.troy".f_remove(3) === "鱼摆摆");
 console.log("f_remove", "鱼摆摆,cn.troy".f_remove(3, 1) === "鱼摆摆cn.troy");
 
-console.log("f_trim", "cn.troy 鱼摆摆 cn.troy".f_trim("cn.troy") === " 鱼摆摆 ");
+console.log(
+  "f_trim",
+  "cn.troy 鱼摆摆 cn.troy".f_trim("cn.troy") === " 鱼摆摆 "
+);
 console.log("f_trim", "   鱼摆摆 ".f_trim() === "鱼摆摆");
 console.log("f_trim", "###鱼摆摆#".f_trim("#") === "鱼摆摆");
 
-console.log("f_trimLeft", "cn.troy 鱼摆摆 cn.troy".f_trimLeft("cn.troy") === " 鱼摆摆 cn.troy");
+console.log(
+  "f_trimLeft",
+  "cn.troy 鱼摆摆 cn.troy".f_trimLeft("cn.troy") === " 鱼摆摆 cn.troy"
+);
 console.log("f_trimLeft", "   鱼摆摆 ".f_trimLeft() === "鱼摆摆 ");
 console.log("f_trimLeft", "###鱼摆摆#".f_trimLeft("#") === "鱼摆摆#");
 
-console.log("f_trimRight", "cn.troy 鱼摆摆 cn.troy".f_trimRight("cn.troy") === "cn.troy 鱼摆摆 ");
+console.log(
+  "f_trimRight",
+  "cn.troy 鱼摆摆 cn.troy".f_trimRight("cn.troy") === "cn.troy 鱼摆摆 "
+);
 console.log("f_trimRight", "   鱼摆摆 ".f_trimRight() === "   鱼摆摆");
 console.log("f_trimRight", "###鱼摆摆#".f_trimRight("#") === "###鱼摆摆");
 
-console.log("f_replace", "鱼摆摆,cn.troy,鱼摆摆".f_replace(",", "&") === "鱼摆摆&cn.troy&鱼摆摆");
+console.log(
+  "f_replace",
+  "鱼摆摆,cn.troy,鱼摆摆".f_replace(",", "&") === "鱼摆摆&cn.troy&鱼摆摆"
+);
 
 console.log("f_startsWith", "鱼摆摆,cn.troy".f_startsWith("鱼摆摆"));
 console.log("f_startsWith", !"鱼摆摆,cn.troy".f_startsWith("cn.troy"));
@@ -54,3 +66,5 @@ console.log("f_isDateTime", "2020-4-12 12:15".f_isDateTime());
 console.log("f_isDateTime", "2020/4/12 12:15:05".f_isDateTime());
 console.log("f_isDateTime", !"2020-4-12 12".f_isDateTime());
 console.log("f_isDateTime", !"鱼摆摆".f_isDateTime());
+
+console.log("isNullOrEmpty", String.f_isNullOrEmpty(null));